struct Overloads moved from Traits to Types namespace
