+++ CMakeLists.txt
@@ -123,7 +123,6 @@ if(HELENA_FLAG_VIEW_HELENA)
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Traits/Function.hpp"
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Traits/Identity.hpp"
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Traits/NameOf.hpp"
-        "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Traits/Overloads.hpp"
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Traits/PowerOf2.hpp"
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Traits/Remove.hpp"
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Traits/SameAll.hpp"
@@ -152,6 +151,7 @@ if(HELENA_FLAG_VIEW_HELENA)
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Types/LocationString.hpp"
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Types/Monostate.hpp"
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Types/Mutex.hpp"
+        "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Types/Overloads.hpp"
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Types/RWLock.hpp"
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Types/ReferencePointer.hpp"
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Types/SourceLocation.hpp"
@@ -173,8 +173,8 @@ if(HELENA_FLAG_VIEW_HELENA)
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Util/Cast.hpp"
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Util/Format.hpp"
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Util/Function.hpp"
-        "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Util/String.hpp"
         "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Util/Sleep.hpp"
+        "${HELENA_PROJECT_DIR}/${HELENA_PROJECT_FRAMEWORK_DIR}/Util/String.hpp"
 	)
 
 	target_sources(Helena INTERFACE ${HELENA_PROJECT_HEADERS})
